pasar

+++ migrations/1_create_beers_table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE beers (
     id int not null auto increment,
     name    varchar(150),
-    price   int(10),
+    price   float(10),
     country varchar(150),
     created_at  DATETIME DEFAULT CURRENT_TIMESTAMP,
     updated_at  DATETIME ON UPDATE CURRENT_TIMESTAMP;
